+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improved youtube id parsing in youtube element
 - Added add-element modal instead of dropdown
 - Added user edit form
+- Added roles and permissions
 
 ## [1.3.10] - 2018-11-22
 ### Added